Provide python 3 compatibility

+++ parler/cache.py
@@ -24,7 +24,7 @@ def get_translation_cache_key(translated_model, master_id, language_code):
     """
     # Always cache the entire object, as this already produces
     # a lot of queries. Don't go for caching individual fields.
-    return 'parler.{0}.{1}.{2}'.format(translated_model.__name__, long(master_id), language_code)
+    return 'parler.{0}.{1}.{2}'.format(translated_model.__name__, int(master_id), language_code)
 
 
 def get_cached_translation(instance, language_code, use_fallback=False):
